refactor(#268): 임시저장 목록 불러오는 로직 및 방식 변경
